Clean up dtos
