+++ src/main/java/com/bompotis/netcheck/service/dto/AbstractPaginatedDto.java
@@ -1,0 +1,38 @@
+package com.bompotis.netcheck.service.dto;
+
+/**
+ * Created by Kyriakos Bompotis on 12/6/20.
+ */
+public abstract class AbstractPaginatedDto {
+
+    private final long totalElements;
+    private final int totalPages;
+    private final int number;
+    private final int size;
+
+    public AbstractPaginatedDto(long totalElements,
+                                int totalPages,
+                                int number,
+                                int size) {
+        this.totalElements = totalElements;
+        this.totalPages = totalPages;
+        this.number = number;
+        this.size = size;
+    }
+
+    public long getTotalElements() {
+        return totalElements;
+    }
+
+    public int getTotalPages() {
+        return totalPages;
+    }
+
+    public int getNumber() {
+        return number;
+    }
+
+    public int getSize() {
+        return size;
+    }
+}
