--- src/main/java/com/bompotis/netcheck/service/dto/PaginatedDomainCheckDto.java
@@ -1,24 +1,0 @@
-package com.bompotis.netcheck.service.dto;
-
-import java.util.List;
-
-/**
- * Created by Kyriakos Bompotis on 12/6/20.
- */
-public class PaginatedDomainCheckDto extends PaginatedDto {
-
-    private final List<DomainCheckDto> domainChecks;
-
-    public PaginatedDomainCheckDto(List<DomainCheckDto> domainChecks,
-                                   long totalElements,
-                                   int totalPages,
-                                   int number,
-                                   int size) {
-        super(totalElements, totalPages, number, size);
-        this.domainChecks = domainChecks;
-    }
-
-    public List<DomainCheckDto> getDomainChecks() {
-        return domainChecks;
-    }
-}
